+++ src/lib.rs
@@ -1,4 +1,5 @@
 pub mod asyncore;
+pub mod http;
 pub mod parser;
 pub mod runtime;
 pub mod template;