+++ src/template/text/mod.rs
@@ -1,0 +1,3 @@
+pub fn t(t: &str) -> String {
+    t.to_string()
+}