+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+### v0.2.0
+
+- Added http module
+- Added rules module
+- Added html module
+- Added text module
+- Added do_html macro
+- Added do_forloop function
+- Added status_map function
+- Improved documentation
+
 ### v0.1.0
 
 - Added asyncore module